+++ src/main/kotlin/appium/recording/Screenshot.kt
