+++ src/main/kotlin/appium/util/EnvironmentConfig.kt
@@ -1,4 +1,4 @@
-package util
+package appium.util
 
 object EnvironmentManager {
