+++ src/main/kotlin/appium/util/PropertiesReader.kt
@@ -1,4 +1,4 @@
-package util
+package appium.util
 
 import java.util.Properties
